fix duplicate number convert
